overwrite fetch globally
