+++ web/src/lib/Utils.ts
@@ -38,3 +38,19 @@ export function isUserModerator(user: User, game: Game) {
             game.moderators.filter((m) => m.id === user.id).length > 0)
     );
 }
+
+/**
+ * Determines the full URL for a request from a relative or absolute path.
+ *
+ * @param path Path to fetch from
+ * @returns The full URL to the API server for the path
+ */
+export function getFullUrl(path: string) {
+    if (path.startsWith('http')) {
+        return path;
+    }
+    if (path.startsWith('/')) {
+        return `${process.env.NEXT_PUBLIC_API_PATH}${path}`;
+    }
+    return `${process.env.NEXT_PUBLIC_API_PATH}/${path}`;
+}
